[CV-410] - User Access Notification Screen
