+++ apps/native/components/FirebaseFunctions.tsx
@@ -1,12 +1,13 @@
 import {doc, getDoc} from "firebase/firestore";
 import {db} from "@native/firebase";
+import AsyncStorage from "@react-native-async-storage/async-storage";
 
 export const fetchUserRole = async (uid: string) => {
     const docRef = doc(db, "users", uid);
     const docSnap = await getDoc(docRef);
 
     if (docSnap.exists()) {
-        console.log("Document data:", docSnap.data());
+        await AsyncStorage.setItem('userUID', docSnap.data().role);
         return docSnap.data().role;
     } else {
         // docSnap.data() will be undefined in this case
@@ -19,6 +20,7 @@ export const fetchUsername = async (uid: string) => {
     const docSnap = await getDoc(docRef);
 
     if (docSnap.exists()) {
+        await AsyncStorage.setItem('userUID', docSnap.data().role);
         console.log("Document data:", docSnap.data());
         return docSnap.data().username;
     } else {
