+++ apps/native/screens/NotificationsManagementScreen.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import {ScrollView, Text, View} from 'react-native';
+
+
+const NotificationsManagementScreen = () => {
+
+    return (
+        <ScrollView>
+            <View>
+                <Text>Request Center</Text>
+
+            </View>
+        </ScrollView>
+    );
+};
+
+export default NotificationsManagementScreen;
